requirements.txt
fastapi==0.86.0
matplotlib==3.5.2
numpy==1.22.4
Pillow==9.3.0
requests==2.27.1
starlette==0.20.4
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
Jinja2==3.1.2
python-multipart==0.0.5

